L3G4200D files

+++ L3G4200D_Lcfg.h
@@ -11,31 +11,31 @@
 #include "Basic_Types.h"
 #include "L3G4200D_Cfg.h"
 /*Full Scale Configuration*/
-#define u8FS_250 0x00
-#define u8FS_500 0x20
-#define u8FS_2000 0x30
-#define u8FULL_SCALE u8FS_250
+#define u8FS_250 ((u8)0x00)
+#define u8FS_500 ((u8)0x20)
+#define u8FS_2000 ((u8)0x30)
+#define u8FULL_SCALE (u8FS_250)
 
 /*Axis Activation Configuration*/
-#define u8AXIS_INACTIVE 0x00
-#define u8X_AXIS_ACTIVE 0x01
-#define u8Y_AXIS_ACTIVE 0x02
-#define u8Z_AXIS_ACTIVE 0x04
+#define u8AXIS_INACTIVE ((u8)0x00)
+#define u8X_AXIS_ACTIVE ((u8)0x01)
+#define u8Y_AXIS_ACTIVE ((u8)0x02)
+#define u8Z_AXIS_ACTIVE ((u8)0x04)
 
 /*High Pass and low Pass filters Configurations*/
-#define u8NO_FILTERS   0x00
-#define u8HPF_ON_DATA  0x11
-#define u8LPF_ON_DATA  0x12
-#define u8BOTH_ON_DATA 0x13
-#define u8HPF_ON_INT   0x14
-#define u8LPF_ON_INT   0x18
-#define u8BOTH_ON_INT  0x1C
+#define u8NO_FILTERS   ((u8)0x00)
+#define u8HPF_ON_DATA  ((u8)0x11)
+#define u8LPF_ON_DATA  ((u8)0x12)
+#define u8BOTH_ON_DATA ((u8)0x13)
+#define u8HPF_ON_INT   ((u8)0x14)
+#define u8LPF_ON_INT   ((u8)0x18)
+#define u8BOTH_ON_INT  ((u8)0x1C)
 
 /*Self Axis movement detection*/
-#define u8INT_DIS 0x00
-#define u8INT1_X  0x03
-#define u8INT1_Y  0x0C
-#define u8INT1_Z  0x30
+#define u8INT_DIS ((u8)0x00)
+#define u8INT1_X  ((u8)0x03)
+#define u8INT1_Y  ((u8)0x0C)
+#define u8INT1_Z  ((u8)0x30)
 
 /*Axis Activation*/
 typedef struct
